+++ ui/src/pages/TableView.vue
@@ -2,12 +2,10 @@
 import { computed, ref, reactive, watch } from 'vue';
 import { computedAsync } from '@vueuse/core';
 import { useApp } from '../app';
-import { model, UiState } from '@platforma-open/milaboratories.table.model';
 import {
-  getAxisId,
   getAxesId,
+  getRawPlatformaInstance,
   type PColumnIdAndSpec,
-  type JoinEntry,
   PTableColumnSpec
 } from '@platforma-sdk/model';
 import * as lodash from 'lodash';
@@ -27,37 +25,19 @@ import {
 } from '@platforma-sdk/ui-vue';
 
 const app = useApp();
-const uiState = app.createUiModel<UiState>(undefined, () => ({
-  settingsOpened: true,
-  filterModel: {},
-  group: {
-    mainColumn: undefined,
-    additionalColumns: [],
-    enrichmentColumns: [],
-    possiblePartitioningAxes: []
-  },
-  partitioningAxes: [],
-  tableState: {
-    gridState: {},
-    pTableParams: {
-      sorting: [],
-      filters: []
-    }
-  }
-}));
 
 /** UI state upgrader */ (() => {
   if ('filtersOpen' in app.model.ui) delete app.model.ui.filtersOpen;
   if (app.model.ui.filterModel === undefined) app.model.ui.filterModel = {};
 })();
 
-const pfDriver = model.pFrameDriver;
+const pfDriver = getRawPlatformaInstance().pFrameDriver;
 const pFrame = computed(() => app.model.outputs.pFrame);
 
 const settingsOpened = computed({
-  get: () => uiState.model.settingsOpened,
+  get: () => app.model.ui.settingsOpened,
   set: (opened) => {
-    uiState.model.settingsOpened = opened;
+    app.model.ui.settingsOpened = opened;
   }
 });
 
@@ -161,7 +141,7 @@ const enrichment = reactive({
 });
 
 watch(
-  () => uiState.model.group,
+  () => app.model.ui.group,
   (group) => {
     const state = [group.mainColumn, group.additionalColumns, group.enrichmentColumns] as const;
     const prevState = [
@@ -177,31 +157,6 @@ watch(
   },
   { immediate: true }
 );
-
-function makeJoin(
-  mainColumn: PColumnIdAndSpec | undefined,
-  additionalColumns: PColumnIdAndSpec[],
-  enrichmentColumns: PColumnIdAndSpec[]
-): JoinEntry<PColumnIdAndSpec> | undefined {
-  if (!mainColumn) return undefined;
-  return {
-    type: 'outer',
-    primary: {
-      type: 'full',
-      entries: [mainColumn, ...additionalColumns].map(
-        (column) =>
-          ({
-            type: 'column',
-            column: column
-          }) as const
-      )
-    },
-    secondary: enrichmentColumns.map((column) => ({
-      type: 'column',
-      column: column
-    }))
-  };
-}
 
 watch(
   () =>
@@ -229,21 +184,10 @@ watch(
     const [_, prevColumn] = prevState;
 
     if (!lodash.isEqual(column, prevColumn)) {
-      const join = makeJoin(column, [], []);
-      uiState.model.group = {
+      app.model.ui.group = {
         mainColumn: column,
         additionalColumns: [],
-        enrichmentColumns: [],
-        possiblePartitioningAxes: (column?.spec.axesSpec ?? []).map(lodash.cloneDeep),
-        join
-      };
-      uiState.model.partitioningAxes = [];
-      uiState.model.tableState = {
-        gridState: {},
-        pTableParams: {
-          sorting: [],
-          filters: []
-        }
+        enrichmentColumns: []
       };
       return;
     }
@@ -268,67 +212,30 @@ watch(
       });
     })();
 
-    const join = makeJoin(column, additional, enrichment);
-    const possiblePartitioningAxes = column.spec.axesSpec.map(lodash.cloneDeep);
-
-    uiState.model.group = {
+    app.model.ui.group = {
       mainColumn: column,
       additionalColumns: additional,
-      enrichmentColumns: enrichment,
-      possiblePartitioningAxes,
-      join
+      enrichmentColumns: enrichment
     };
   }
 );
 
-const partitioningPlaceholder = computed(() =>
-  !column.selected ? 'First, select the main column' : 'Select axes for partitioning'
-);
-const partitioningOptions = computed(() => {
-  return uiState.model.group.possiblePartitioningAxes.map((axis, i) => ({
-    text: axis.annotations?.['pl7.app/label']?.trim() ?? 'Unlabelled axis ' + i.toString(),
-    value: getAxisId(axis)
-  }));
-});
-const partitioningSelected = computed({
-  get: () => uiState.model.partitioningAxes,
-  set: (partitioningAxes) => {
-    const options = partitioningOptions.value;
-    partitioningAxes.sort((lhs, rhs) => {
-      const li = lodash.findIndex(options, (option) => lodash.isEqual(option.value, lhs));
-      const ri = lodash.findIndex(options, (option) => lodash.isEqual(option.value, rhs));
-      return li - ri;
-    });
-
-    uiState.model.partitioningAxes = partitioningAxes;
-  }
-});
-const partitioningDisabled = computed(() => !column.selected);
-const partitioning = reactive({
-  label: 'Axes for partitioning',
-  placeholder: partitioningPlaceholder,
-  options: partitioningOptions,
-  selected: partitioningSelected,
-  disabled: partitioningDisabled
-});
-
 const tableState = computed({
-  get: () => uiState.model.tableState,
+  get: () => app.model.ui.tableState,
   set: (tableState) => {
-    if (!lodash.isEqual(tableState, uiState.model.tableState)) {
-      uiState.model.tableState = tableState;
+    if (!lodash.isEqual(tableState, app.model.ui.tableState)) {
+      app.model.ui.tableState = tableState;
     }
   }
 });
-const tableSettings = computed(
-  () =>
-    ({
-      sourceType: 'pframe',
-      pFrame: app.model.outputs.pFrame,
-      join: uiState.model.group.join,
-      sheetAxes: uiState.model.partitioningAxes,
-      pTable: app.model.outputs.pTable
-    }) satisfies PlDataTableSettings
+const tableSettings = computed<PlDataTableSettings | undefined>(() =>
+  app.model.ui.group.mainColumn
+    ? {
+        sourceType: 'ptable',
+        pTable: app.model.outputs.pTable,
+        sheets: app.model.outputs.sheets
+      }
+    : undefined
 );
 const columns = ref<PTableColumnSpec[]>([]);
 const tableInstance = ref<PlAgDataTableController>();
@@ -394,14 +301,6 @@ const tableInstance = ref<PlAgDataTableController>();
       v-model="enrichment.selected"
       clearable
       :disabled="enrichment.disabled"
-    />
-    <PlDropdownMulti
-      :label="partitioning.label"
-      :placeholder="partitioning.placeholder"
-      :options="partitioning.options"
-      v-model="partitioning.selected"
-      clearable
-      :disabled="partitioning.disabled"
     />
   </PlSlideModal>
 </template>